Move responsibility for timers into ElectionTiming.java
